created view post and new post modules to clean up server.js
